Fixed DeltaTime Begin not float

+++ Brawler-server/Server/Server.cs
@@ -130,7 +130,7 @@ namespace BrawlerServer.Server
         public bool IsRunning { get; set; }
         public uint Time { get; private set; }
         // does NOT count looptime
-        public uint DeltaTime { get; private set; }
+        public float DeltaTime { get; private set; }
 
         public uint MaxIdleTimeout { get; private set; }
